match datatype with internal type for object/array id

+++ vistle/core/shm.cpp
@@ -229,20 +229,20 @@ void Shm::unlockDictionary() const {
 #endif
 }
 
-size_t Shm::objectID() const {
+int Shm::objectID() const {
     return m_objectId;
 }
 
-size_t Shm::arrayID() const {
+int Shm::arrayID() const {
     return m_arrayId;
 }
 
-void Shm::setObjectID(size_t id) {
+void Shm::setObjectID(int id) {
     vassert(id > m_objectId);
     m_objectId = id;
 }
 
-void Shm::setArrayID(size_t id) {
+void Shm::setArrayID(int id) {
     vassert(id > m_arrayId);
     m_arrayId = id;
 }
